Added Fragment, and navigation graph in the project.
