+++ app/src/main/java/com/puzzlebench/myanimereviewk/presentation/AnimeDetailFragment.kt
@@ -1,0 +1,16 @@
+package com.puzzlebench.myanimereviewk.presentation
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.puzzlebench.myanimereviewk.R
+
+class AnimeDetailFragment : Fragment() {
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
+                              savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.fragment_anime_detail, container, false)
+    }
+}
