+++ app/src/main/java/com/puzzlebench/myanimereviewk/presentation/MainActivity.kt
@@ -2,35 +2,23 @@ package com.puzzlebench.myanimereviewk.presentation
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.navigation.fragment.NavHostFragment
 import com.puzzlebench.myanimereviewk.R
 import com.puzzlebench.myanimereviewk.presentation.states.AnimeListState
+import com.puzzlebench.myanimereviewk.presentation.viewModels.AnimeListViewModel
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class MainActivity : AppCompatActivity() {
-
-    private val animeListViewModel: AnimeListViewModel by viewModel()
+class MainActivity : AppCompatActivity(R.layout.activity_main) {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        initObservers()
-        animeListViewModel.getAnimeTopList()
+        initView()
     }
 
-    private fun initObservers() {
-        animeListViewModel.getAnimeListStateLiveData.observe(this, { state ->
-            when (state) {
-                is AnimeListState.Loading -> {
-                    //load
-                }
-                is AnimeListState.Success -> {
-                    //data
-                }
-                else -> {
-                    //error
-                }
-            }
-        })
+    private fun initView() {
+        val navController =
+            supportFragmentManager
+                .findFragmentById(R.id.fraNavigationAnimeReview) as NavHostFragment
+        navController.navController
     }
 }
