+++ app/src/main/java/com/puzzlebench/myanimereviewk/presentation/di/ViewModelModule.kt
@@ -1,8 +1,8 @@
 package com.example.myitemssearchk.presentation.di
 
 import androidx.lifecycle.MutableLiveData
-import com.puzzlebench.myanimereviewk.presentation.AnimeDetailViewModel
-import com.puzzlebench.myanimereviewk.presentation.AnimeListViewModel
+import com.puzzlebench.myanimereviewk.presentation.viewModels.AnimeDetailViewModel
+import com.puzzlebench.myanimereviewk.presentation.viewModels.AnimeListViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
